requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
python-frontmatter==1.1.0 